Create and style Counter and Todo components
Set up the framework for Counter and Todo components in the application, enabling the Counter to increase and decrease a count. The Todo component has been equipped with the ability to add and complete tasks. As of now, the Todo component makes use of static data from data.json. The overall appearance is handled by the newly added styles in index.css, which also pertain to these new components.